If no sub is mentioned in unsubscribe show all subscriptions (something epic)

+++ cogs/reddit.py
@@ -3,7 +3,6 @@ import random
 from discord.ext import commands, tasks
 import asyncpraw as praw
 import asyncprawcore as prawcore
-
 import functions.database as db
 from functions.style import Emotes, Colours, TIME
 from Nix import CLIENT_ID, SECRET_KEY, USER_AGENT
@@ -46,15 +45,16 @@ class Reddit(commands.Cog):
 
     @commands.slash_command(name='unsubscribe', description="Unsubscribe to daily posts from the given subreddit")
     @discord.commands.default_permissions(manage_guild=True)
-    async def unsubscribe_from_sub(self, ctx, sub):
-        # TODO should output current subs if sub is not provided
-
-        if sub not in db.single_SQL("SELECT Subreddit FROM Subreddits WHERE GuildID=%s", (ctx.guild_id,)):
-            await ctx.respond("This server is not subscriped to r/{0} {1}".format(sub, Emotes.SUPRISE))
-            return
-        db.single_SQL("DELETE FROM Subreddits WHERE GuildID=%s AND Subreddit=%s ",
-                      (ctx.guild_id, sub))  # Delete the subscription out of the SQL
-        await ctx.respond("This server is now unsubscribed from {0} {1}".format(sub, Emotes.SNEAKY))
+    async def unsubscribe_from_sub(self, ctx, sub: discord.Option(required=False)):
+        if not sub:
+            await ctx.respond(await self.getSubs(ctx))
+        else:
+            if sub not in db.single_SQL("SELECT Subreddit FROM Subreddits WHERE GuildID=%s", (ctx.guild_id,)):
+                await ctx.respond("This server is not subscribed to r/{0} {1}".format(sub, Emotes.SUPRISE))
+            else:
+                db.single_SQL("DELETE FROM Subreddits WHERE GuildID=%s AND Subreddit=%s ",
+                              (ctx.guild_id, sub))  # Delete the subscription out of the SQL
+                await ctx.respond("This server is now unsubscribed from {0} {1}".format(sub, Emotes.SNEAKY))
 
     @commands.slash_command(name='subscriptions', description="Get a list of the subscriptions of the server")
     async def getSubs(self, ctx):
